requirements.txt
certifi==2017.11.5
chardet==3.0.4
future==0.16.0
google-api-python-client==1.6.4
httplib2==0.10.3
idna==2.6
oauth2client==4.1.2
oauthlib==2.0.6
pyasn1==0.4.2
pyasn1-modules==0.2.1
python-twitter==3.3
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
